consensus_inner

Finds the consensus risk haplotype shared between case subjects (main body of algorithm for get.consensus function).

consensus_inner(direction = c("right", "left"), consensus, haps, riskhaps,
  site.start, site.end, groups, max.leave, verbose)

Arguments

direction
Which direction to move from site of interest. Options are "right" (genomic positions higher than site of interest) and "left" (genomic positions lower than site of interest).

consensus
Skeleton data frame to which consensus haplotype segments will be added

haps
Matrix of phased haplotypes

riskhaps
Vector of identifiers for risk haplotypes

site.start
Genomic position at start of site of interest

site.end
Genomic position at end of site of interest

groups
Internal variable tracking which subjects are in the current haplotype group

max.leave
Maximum number of haplotypes that may leave a haplotype group without starting their own new consensus haplotype segment

verbose
Logical indicating whether progress updates should be printed. Default FALSE.

Value

List containing data frame of consensus haplotype segments and vector of split positions.