plot_consensus

Produces a plot of consensus haplotypes. Haplotype segments belonging to a single group are identified by color and (optionally) symbol. Site of interest may be plotted; start and end locations appear as solid lines. Genomic positions where consensus haplotype splits into subgroups may also be plotted; these appear as dotted lines. Subjects may be ordered by group membership with group names included alongside the subject identifiers.

plot_consensus(consensus, splits = NULL, site.start = NULL,
  site.end = NULL, colors = NULL, ncolors = 9, symbols = TRUE,
  groups = NULL, group.names = NULL, label.size = 1, title.size = 1)

Arguments

consensus
Data frame containing consensus haplotype segments (columns are riskhap ID, segment start, segment end, and consensus haplotype segment ID. Output from get.consensus().

splits
Vector of genomic positions where consensus haplotype splits into groups. Default NULL; if included, dashed lines will be drawn at each split position. Output frum get.consensus().

site.start
Genomic position of start of site of interest. Default NULL; if included, a solid line will be drawn at site.start.

site.end
Genomic position of end of site of interest. Default NULL; if included, a solid line will be drawn at site.end.

colors
Vector of colors for haplotype segments. Default NULL (program will generate colors).

ncolors
Number of unique colors to include on plot. Max 9, default 9. Colors are generated from brewer.pal to be as easily distinguishable as possible and repeated if more haplotype segments are present than colors.

symbols
Logical indicating whether symbols should be overlaid line segments for each consensus haplotype segment. Default TRUE.

groups
Vector identifying group membership for each subject. Default NULL.

group.names
Names of subgroups of subjects. Default NULL.

label.size
Controls relative size of subject and group labels. Default 1.

title.size
Controls relative size of text in title and axis labels. Default 1.

Value

Creates plot